+++ src/App.tsx
@@ -9,8 +9,7 @@ import LoginForm from "./pages/user/LoginForm"; // 로그인폼
 import HjWebTest from "./pages/main/HjWebTest";
 
 import PlanFilterSelector from "./pages/plan/PlanFilterSelector"; // 계획 필터 입력
-import PlanList from "./pages/plan/PlanList"; // 일정 리스트
-import PlanModify from "./pages/plan/PlanModify"; // 일정 수정
+import Plan from "./pages/plan/Plan";
 
 function App() {
   return (
@@ -25,8 +24,8 @@ function App() {
               path="/plan/filter/selector"
               element={<PlanFilterSelector />}
             />
-            <Route path="/plan/list" element={<PlanList />} />
-            <Route path="/plan/modify" element={<PlanModify />} />
+            <Route path="/plan/list" element={<Plan />} />
+            {/* <Route path="/plan/modify" element={<PlanModify />} /> */}
           </Routes>
         </Layout>
       </BrowserRouter>
